todo db connected
