+++ package.json
@@ -2,8 +2,10 @@
   "name": "rajat_ldp_node_ts",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "src/index.ts",
   "scripts": {
+    "build": "tsc --project ./",
+    "start": "ts-node-dev --respawn --clear src/app.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -18,13 +20,14 @@
   "homepage": "https://github.com/zemoso-int/rajat_Ldp_node_TS#readme",
   "dependencies": {
     "body-parser": "^1.20.0",
-    "dotenv": "^16.0.0",
     "express": "^4.18.1",
-    "mongodb": "^4.5.0",
-    "typescript": "^4.6.4"
+    "mongodb": "^4.5.0"
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
-    "@types/node": "^17.0.31"
+    "@types/node": "^17.0.31",
+    "dotenv": "^16.0.0",
+    "ts-node-dev": "^1.1.8",
+    "typescript": "^4.6.4"
   }
 }
